kites/metrics: rename queue to storage
